New wallet screen design

Closes #30

See merge request irma/irmamobile!67

+++ lib/src/screens/enrollment/widgets/introduction.dart
@@ -62,7 +62,7 @@ class _IntroductionState extends State<Introduction> {
               children: <Widget>[
                 DotsIndicator(
                   dotsCount: pageLength,
-                  position: currentIndexPage,
+                  position: currentIndexPage.toDouble(),
                   decorator: DotsDecorator(
                     color: Colors.grey[400],
                     activeColor: Colors.grey[700],
